requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0
accelerate>=0.20.0

# Web interface
streamlit>=1.25.0

# Configuration and utilities
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional dependencies for enhanced functionality
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# For better performance (optional)
# torch-audio>=2.0.0  # Uncomment if using audio features
# sentence-transformers>=2.2.0  # Uncomment for semantic similarity
